Transliteration of CityHash code from C++ to Go.
This is a straightforward implementation in Go of the CityHash
non-cryptographic hashing algorithm, done by transliterating the C++
implementation.  The 32- and 64-bit functions are implemented, but not the
128-bit functions.

The unit tests were constructed by extracting the test vectors from the C++
unit test file.  The convert_tests script does this extraction.