Update convert_tests to work with Python3.
This mostly amounts to working around Python 3's aggressive idiocy regarding
the separation of "encoded" vs. "decoded" string values. It's not enough to
decode on input: Once you have text, you have to manually push the encoding
rule down into the output stream, or -- as I did here -- buffer up the whole
thing and waste a bunch of memory at the end.

There can be no sensible reason for simple things to be this complicated.